requirements.txt
beeai-framework
uvicorn
python-dotenv
requests
fastapi
pydantic